+++ rust/theoros/src/handlers/mod.rs
@@ -1,3 +1,2 @@
-pub mod build_calldata;
 pub mod rest;
 pub mod websocket;
